trash functionality for collections
Deleted collections go to trash, similar to items and show up
among them. Subcollections of a deleted collection are also trashed
though only the parent collection is displayed in trash itemTree.

When a parent collection is permanently deleted or restored,
same applies to subcollections.

Items are not affected by trashing/restoration/permanent deletion
of collections.