accessibility improvements for quick format dialog
Keyboard navigation and screen reader accessibility for VoiceOver and
NVDA.

1. When the cursor is moved, the selected bubble is focused so that
screen readers announce the name.
2. _getSelectedBubble will take into account currently focused bubble.
It prevents CMD-ArrowDown from opening citation-properties for a bubble
different from the one that was announced.
3. Added explicit keyboard navigation within reference panel.
A selected reference will always be focused, and a focused reference is
always selected.
By default, VoiceOver will focus but not select, and NVDA will allow
selection but without focus won't read it.
4. Keyboard navigation and aria labels for citation-properties.
5. Shift-tab from the input field will move focus to the Zotero-icon
menu.
6. Shift-tab from the reference panel moves focus back to the input
field.
7. Tab from outside of the reference panel will focus the first bubble.
8. Space on the bubble will move focus from the bubble onto the editor
and place cursor after the bubble. Required to force VoiceOver to let go
of the bubble and allow to type after it.