aria description with instructions
Added aria-description to bubble, input and item to tell the user
what keypresses are available. Works with VoiceOver and NVDA,
JAWS refuses to read it (possibly related to https://github.com/FreedomScientific/standards-support/issues/719#)
Also, bubble had to be made a button to have the right semantics for
aria-description to be read, and JAWS by default instructs to press Enter
to open it which is not how the dialog works.