more dialog improvements
1. On arrowLeft from the input, move focus to the bubble on the left
if the cursor is in the very beginning of the input. Similarly, on
arrowRight, move focus to the next bubble if the cursor is in the end
of the input.
2. Hide reference panel when citation properties popup is opened.
When the popup is closed, show reference panel again and move focus
back to the input if it is opened. Otherwise, focus the bubble.
3. Throttle Enter and Escape keypress events that can close the dialog.