simplified _resize that works better
Simplified _resize just calls window.resizeTo and sets the height to
be the one of the top-most component. This works better since it
allows the height to be shrunk if the entire line of bubbles is erased.

Some tweaks to drag-drop behavior related to resizing. Deleting the input
immediately on blur interrupts dragstart event for some mozilla related
reasons. As a temp solution, added a half-second delay which should be
replaced for something better.