fix editable-text 'undefined' on escape in reader
When reader is opened, Escape is picked up in capturing mode and it
refocuses the scrollable section of the reader. If any editable-text
is focused, its 'blur' handler will fire before 'keypress' handler.
That means that when editable-text's escape handler sets the input's
back to it's initial value, dataset.initialValue had already been cleared, so
input's value is 'undefined'.

A timeout for reader.focus() allows editable-text 'keypress' handler
to revert back to initial value before it's removed.