non-modal dialog windows
voiceover can't seem to get inside of dialog windows
opened as a modal. It remains on the main window even
though the dialog window is focused. Voiceover has its own particular way
of handling the cursor and it does not follow the
focus as closely as NVDA or JAWS. This is an issue for VPAT.

Interestingly, it only happens if a window is openend as a modal. To avoid
redoing the entire logic surrounding dialog windows, maybe we can:
1. make the windows non-modals
2. pass a callback function as a parameter into the dialog window
that will run when the dialog is accepted
3. to make it feel at least somewhat similar to previous blocking modal,
we can cover the entire screen with an opaque component that will also
consume all click events until the dialog is closed.

This is a sketch of what it could look like specifically for "New collection"
dialog.