proof of concept: display dialogs within <browser>
As a potential solution to the voiceover not going
inside of dialogs opened via window.openDialog with
'modal' as an option, we can render the dialog inside
of a <browser> component and display if inside of a
wrapper dialog with .showModal. It does require some
refactoring but the look and feel remains about the same
as before. This a simpler replica of what seems to
happen with Firefox on 'about:support' > "Refresh firefox".

This is a proof of concept to see how it could work
specifically with "New collection" dialog, so there's
some hardcoding and messiness that will need to be
cleaned up and generalized - but the main idea
should be clear.