part vpat 16: menu to move collections
- added context menu to move a collection into another
collection. Since one can drag a collection into another
library, the top-level entries of the contextmenu are
libraries and not collections as with itemTree.
- refactoring of collectionTree to pull out copyCollection()
and copyItems() from out of the scope of CollectionTree.onDrop.
If a collection is moved into another library, the functionality
of copying over a collection needs to be exposed to ZoteroPane
without the actual drag-drop happening.