vpat: hack for win10 JAWS to read text in settings
Turns out JAWS on windows 10 will not read the headers and
descriptions when focus enters the groups (though it works
just fine on win11). As a, hopefully, temporary workaround
we can fetch aria-labels, descriptions and headers and
announce them with an alert when focus moves between
groups. This applies exclusively to windows 10.