better separation of responsibilities
General refactoring for better separation of responsibilities
between classes and components.

- Replace Citation singleton that handled both the items
storage AND some handling of IO from bubble-input
with two singletons: CitatitonIOManager for IO and
CitationDataManager just for handling citation items storage.
- CitationDataManager.items stores both Zotero.Items and citationItems
to not have to convert them back and forth in bubbleInput.
- Move the logic of arranging and rendering bubbles to bubbleInput -
citationDialog just passes a list of citation items
to it, and bubbleInput is responsibile for everything else.
Thus, some Helpers functions (building bubble string) were moved
to bubbleInput, since now it's the only place it is used in.
- Refactor bubbleInput's to not insert <input> between bubbles
on click or during arrow navigation. Instead, when a bubble is added,
always surround it with <input>s. That way, arrowRight/left navigation
in bubbleInput is handled by KeyboardHandler, same as it is done in the
rest of citationDialog.
- Move handling of top-level keypresses into CitatitonIOManager. This is
where all general keypresses are handled, and bubbleInput handles or
keypresses that are specific to bubbles (e.g. backspace to delete bubbles).
- Refactoring to not pass any functions or Helpers into initialization
of classes and keep them independent of each other.

Also, some more general improvements.
- do not allow duplicate items to be added. Exclude them from search results.
- add +/- column to itemTree in library mode to add or remove an item
on click.
- remove unused settings checkboxes for now. keepSorted will resort
citations if checked after being un-checked