reworked selected/focused/pre-selected logic
- removed pre-selected state
- in list mode, keep focus in the bubble-input while
also managing "selected" state in items list. After something
is typed, the first item is marked as selected,
arrow-down will select the next item, and arrow-up will select
the previous item. Focus remains in the input. One can also tab out of the
bubble-input to focus the selected item in the list and
still navigate the list with arrows.
- in library mode, the first item in suggested items is also
marked as selected and Enter from input will add it.
Since the layout in this case is horizontal, and arrow-right/left
are used to navigate within bubble-input, one has to press arrow-down
to move focus to the first item before navigating suggested items
with left-right arrows.
- to avoid confusion as to what item is currently selected after
searching, any collapsed sections will be expanded when something is
typed.
- while multi-selecting items via shift-arrow in library mode,
if the next node is a collapsed deck, it will be expanded and
all of its items will be selected. That is to avoid adding another
state for what a "selected" deck would look and behave like, and seems
consistent with expanding the deck automatically when search runs.