try to keep focus in main components after clicks
Every node that is focusable will receive focus on click.
In many cases, we don't want this behavior. For example,
if you expand a collapsed deck of selected items,
focus should ideally stay in bubble-input. With this,
we keep track of the last focused node before a click
happened and return focus to it after click happened.
Similarly, return focus to last focused node before a
popup was opened, when it is hidden.