better align first bubbles on each line
- set initial width on all inputs as just 0
- on focus of an empty input, set min-width to make
the cursor visible and add a negative end-margin to not
shift bubbles

That way, the first bubble on each line has no offset.
It does mean that the input at the start of each line
is slightly misaligned with bubbles, since we do not reserve
space for it, but it still seems to look better.