remove accept button from taborder
So that one has one less tab to press to get to the
content. It may cause some confusion from a11y perspective
but QF does not have a focusable Accept button, so this would
be consistent with that. And, since Enter does the same thing,
the vpat should be ok