refactoring of the fade effect
To not overlap with scrollbar. Use absolutely positioned
linear gradient instead of mask whose height does not
go all the way down to the scrollbar.

If the scrollbar is hovered, its height increases on mac
and overlaps with the items. So to avoid gap overlapping
with the fade effect in that case, the gap between suggested
items and the tables is also slightly increased