edit bibliography dialog redesign
- updated the dialog's layout to remove empty space
around the edges
- replaced the references list with a proper itemTree
- removed selectItemsDialog.js as a dependency.
Some behaviors of selectItemsDialog.js conflict with
what needs to happen in editBibliographyDialog. For
example, custom logic is needed to save column preferences
of two itemTrees in the same window. Also, onActivate
handler of itemTree does not need to accept the entire
bibliography dialog, even though it makes sense when
a single item is being selected.
- move handling of clickable buttons from citationDialog
into itemTree. It allows us to also add +/- buttons
to itemTrees in this dialog.
- library itemTree has a column with + button to add
the item into the list of references. It disappears
if the item is already included
- references itemTree has a column with - button to
remove the item from the list of references. If the
item is not cited, it is removed from this table. If
the item is cited, it will remain but will become
faded. It's column button will change to +, which will
re-add that item into bibliography.
- references itemTree also has a "Edited" column which
indicates of the citation has been manually changed by the
user or not.
- there is a splitter between library trees and the
references itemTree. It can be dragged to make
more space for references and can be collapsed to hide
library trees completely.

Additional minor fixes:
- do not display tooltip when a button in virtualized-table
is hovered over
- cleanup of integration.css to remove unused styles
from editBibliographyDialog and citationDialog