search handler support for inserting annotations
- when adding selected/open/cited items only keep
items that have annotations
- when searching the library, annotations can be
found by searching for the annotation text/comment
itself or its top-level item. For consistency, same
applies to the notes as well.
- _ensureRelevantItemsAreLoaded is used to load
all ancestors/descendants/siblings of search matches
to avoid "Item X is not loaded" errors.
- for the above, allow Zotero.Items.getAnnotations to return only
annotation IDs due to an edge cases when the
._annotation cache would be set on the attachment but
the annotations themselves would not be loaded, which
errored on Zotero.Items.get(annotationIDs).