citation dlg: UI to insert annotations
In both modes:
- one can select multiple items, when the dialog is accepted,
all of them are inserted
- a regular item or attachment can be selected, in which
case all annotations of the item will be inserted

Library mode:
- +/- icons do not appear in annotations mode
- annotation cards are slightly larger than item
cards of regular items to allow for an extra line
of text (text, comment, parent item's title)

List mode:
- annotations appear as child nodes within their parent
top-level item. The top-level item can be expanded
or collapsed by clicking on the twisty or on right/left
keypress when it is focused

Also:
- slightly reworked expand/collapse height logic to
rely on max-height instead of height. This allows
us to collapse a top-level item and have the parent
section shrink instead of holding on to the original height