potential fix to CI tagsbox test timeout
Zotero.Utilities.Internal.activate was always awaiting
for the "load" event before pulling the window to the front
on linux x11. That means that if the window was already
open (as during test runs), no activation would ever happen,
and the tagsbox test would timeout.
With this change, _X11BringToForeground will be called immediately
if the document's readyState is "complete".